refactor(context): streamline ResolverPayload and update imports
- Removed redundant definitions of `ResolverPayload` and consolidated its import across relevant files for better organization.
- Introduced a new constants file to define `DERIVED_DEPENDENCIES`, enhancing clarity in dependency management.
- Updated various files to reflect the changes in imports and ensure consistency in type definitions.
