+++ packages/core/src/utils/constants.ts
@@ -1,0 +1,1 @@
+export const DERIVED_DEPENDENCIES = "loom.derived-from-dependencies"
